refactor(core): move formatBandLabel to @oga/core (#207)
Pure display formatter for ApproachBandStat ranges. Belongs in @oga/core
alongside the ApproachBandStat type definition — no platform dependencies,
reusable across web and mobile.

Lives in stats.ts (not units.ts) because it consumes ApproachBandStat and
units.ts → stats.ts would invert the existing dep direction.

- Add formatBandLabel + tests to @oga/core
- Strip from apps/web/src/pages/stats/format.ts
- Update StatTiles.tsx import to @oga/core

Closes #207
